requirements.txt
flask==2.2.5
flask-cors==3.0.10
werkzeug==2.2.3
python-dotenv==0.19.0
requests==2.26.0
mysql-connector-python==8.0.32
openai
google-cloud-speech
pydantic
PyJWT==2.6.0